fix float

+++ multi-gammon-core/src/main/java/org/akoshterek/backgammon/agent/HeuristicAgent.scala
@@ -22,34 +22,34 @@ class HeuristicAgent(override val path: Path) extends AbsAgent("Heuristic", path
 
   def evalContact(board: Board): Reward = {
     val reward = Reward.rewardArray[Float]
-    reward(Constants.OUTPUT_WIN) = evaluate(board.anBoard(Board.SELF)).toFloat
+    reward(Constants.OUTPUT_WIN) = evaluate(board.anBoard(Board.SELF))
     new Reward(reward)
   }
 
-  private def evaluate(points: Array[Int]): Double = {
-    var equity: Double = 0.0
+  private def evaluate(points: Array[Int]): Float = {
+    var equity: Float = 0.0f
     val atHome: Int = 15 - points.sum
 
     // 1/15th of a point per man home
-    equity += atHome / 15.0
+    equity += atHome / 15.0f
 
     // -1/5th of a point per man on the bar
-    equity -= points(Board.BAR) / 5.0
+    equity -= points(Board.BAR) / 5.0f
 
     for (i <- 0 until Board.HALF_BOARD_SIZE - 1) {
       // -1/10th of a point for each blot
       if (points(i) == 1) {
-        equity -= 0.10
+        equity -= 0.10f
       }
       // +1/20th for contiguous points
       else if (i > 0 && points(i) >= 2 && points(i - 1) >= 2) {
-        equity += 0.05
+        equity += 0.05f
       }
 
       val dist: Int = 25 - i
 
       // value based on closeness to home
-      equity += (12.5 - dist) * points(i).toDouble / 225.0
+      equity += (12.5f - dist) * points(i).toFloat / 225.0f
     }
 
     equity
